[TravelSet-13] Make category implementation and add other entities, services, controllers
