+++ src/main/java/com/travelset/laba3/exceptions/EntityElementAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.travelset.laba3.exceptions;
+
+public class EntityElementAlreadyExistsException extends RuntimeException {
+    public EntityElementAlreadyExistsException(Class entityClass, Integer id) {
+        super(entityClass.getName() + " with id=" + id + " Already Exists.");
+    }
+}
