+++ src/main/java/com/travelset/laba3/model/repository/CategoryRepository.java
@@ -1,0 +1,12 @@
+package com.travelset.laba3.model.repository;
+
+import com.travelset.laba3.model.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+    Category findFirstByOrderByIdAsc();
+    boolean existsCategoryById(int id);
+
+}
