+++ src/main/java/com/travelset/laba3/model/repository/OrderRepository.java
@@ -1,0 +1,10 @@
+package com.travelset.laba3.model.repository;
+
+import com.travelset.laba3.model.entity.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+    Order findFirstByOrderByIdAsc();
+}
