+++ src/main/java/com/travelset/laba3/model/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.travelset.laba3.model.repository;
+
+import com.travelset.laba3.model.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
+    User findFirstByOrderByIdAsc();
+}
