+++ src/main/java/com/travelset/laba3/services/OrderItemService.java
@@ -1,0 +1,8 @@
+package com.travelset.laba3.services;
+
+import com.travelset.laba3.model.entity.OrderItem;
+
+public interface OrderItemService {
+    OrderItem getOrderItem(int id);
+    void createOrUpdateOrderItem(int orderId, int itemId, int quantity);
+}
