--- src/main/java/com/travelset/model/entity/Backpack.java
@@ -1,103 +1,0 @@
-package com.travelset.model.entity;
-
-import javax.persistence.*;
-import java.util.Collection;
-import java.util.Objects;
-
-@Entity
-@Table(name = "backpacks")
-public class Backpack {
-    private int id;
-    private String description;
-    private int weight;
-    private int quantity;
-    private double maxCapacity;
-    private Collection<Order> orders;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "Id")
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    @Basic
-    @Column(name = "Description")
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    @Basic
-    @Column(name = "Weight")
-    public int getWeight() {
-        return weight;
-    }
-
-    public void setWeight(int weight) {
-        this.weight = weight;
-    }
-
-    @Basic
-    @Column(name = "Quantity")
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    @Basic
-    @Column(name = "MaxCapacity")
-    public double getMaxCapacity() {
-        return maxCapacity;
-    }
-
-    public void setMaxCapacity(double maxCapacity) {
-        this.maxCapacity = maxCapacity;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        Backpack backpack = (Backpack) o;
-
-        if (id != backpack.id) return false;
-        if (weight != backpack.weight) return false;
-        if (quantity != backpack.quantity) return false;
-        if (Double.compare(backpack.maxCapacity, maxCapacity) != 0) return false;
-        return Objects.equals(description, backpack.description);
-    }
-
-    @Override
-    public int hashCode() {
-        int result;
-        long temp;
-        result = id;
-        result = 31 * result + (description != null ? description.hashCode() : 0);
-        result = 31 * result + weight;
-        result = 31 * result + quantity;
-        temp = Double.doubleToLongBits(maxCapacity);
-        result = 31 * result + (int) (temp ^ (temp >>> 32));
-        return result;
-    }
-
-    @OneToMany(mappedBy = "status")
-    public Collection<Order> getOrders() {
-        return orders;
-    }
-
-    public void setOrders(Collection<Order> orders) {
-        this.orders = orders;
-    }
-}
