--- src/main/java/com/travelset/model/entity/Category.java
@@ -1,61 +1,0 @@
-package com.travelset.model.entity;
-
-import javax.persistence.*;
-import java.util.Collection;
-import java.util.Objects;
-
-@Entity
-@Table(name = "categories")
-public class Category {
-    private int id;
-    private String name;
-    private Collection<Itemtype> itemtypes;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "Id")
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    @Basic
-    @Column(name = "Name")
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        Category that = (Category) o;
-
-        if (id != that.id) return false;
-        return Objects.equals(name, that.name);
-    }
-
-    @Override
-    public int hashCode() {
-        int result = id;
-        result = 31 * result + (name != null ? name.hashCode() : 0);
-        return result;
-    }
-
-    @OneToMany(mappedBy = "category")
-    public Collection<Itemtype> getItemtypes() {
-        return itemtypes;
-    }
-
-    public void setItemtypes(Collection<Itemtype> itemtypes) {
-        this.itemtypes = itemtypes;
-    }
-}
