--- src/main/java/com/travelset/model/entity/Itemtype.java
@@ -1,72 +1,0 @@
-package com.travelset.model.entity;
-
-import javax.persistence.*;
-import java.util.Collection;
-import java.util.Objects;
-
-@Entity
-@Table(name = "itemtypes")
-public class Itemtype {
-    private int id;
-    private String name;
-    private Collection<Item> items;
-    private Category category;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "Id")
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    @Basic
-    @Column(name = "Name")
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        Itemtype itemtype = (Itemtype) o;
-
-        if (id != itemtype.id) return false;
-        return Objects.equals(name, itemtype.name);
-    }
-
-    @Override
-    public int hashCode() {
-        int result = id;
-        result = 31 * result + (name != null ? name.hashCode() : 0);
-        return result;
-    }
-
-    @OneToMany(mappedBy = "itemtype")
-    public Collection<Item> getItems() {
-        return items;
-    }
-
-    public void setItems(Collection<Item> items) {
-        this.items = items;
-    }
-
-    @ManyToOne
-    @JoinColumn(name = "CategoryId", referencedColumnName = "Id", nullable = false)
-    public Category getCategory() {
-        return category;
-    }
-
-    public void setCategory(Category category) {
-        this.category = category;
-    }
-}
