--- src/main/java/com/travelset/model/entity/Orderline.java
@@ -1,73 +1,0 @@
-package com.travelset.model.entity;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "orderlines")
-public class Orderline {
-    private int id;
-    private int quantity;
-    private Item item;
-    private Order order;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "Id")
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    @Basic
-    @Column(name = "Quantity")
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        Orderline that = (Orderline) o;
-
-        if (id != that.id) return false;
-        if (quantity != that.quantity) return false;
-
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        int result = id;
-        result = 31 * result + quantity;
-        return result;
-    }
-
-    @ManyToOne
-    @JoinColumn(name = "ItemId", referencedColumnName = "Id", nullable = false)
-    public Item getItem() {
-        return item;
-    }
-
-    public void setItem(Item item) {
-        this.item = item;
-    }
-
-    @ManyToOne
-    @JoinColumn(name = "OrderId", referencedColumnName = "Id", nullable = false)
-    public Order getOrder() {
-        return order;
-    }
-
-    public void setOrder(Order order) {
-        this.order = order;
-    }
-}
