--- src/main/java/com/travelset/model/repository/BackpackRepository.java
@@ -1,10 +1,0 @@
-package com.travelset.model.repository;
-
-import com.travelset.model.entity.Backpack;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BackpackRepository extends JpaRepository<Backpack, Integer> {
-
-}
