--- src/main/java/com/travelset/model/repository/CategoryRepository.java
@@ -1,10 +1,0 @@
-package com.travelset.model.repository;
-
-import com.travelset.model.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-
-}
