--- src/main/java/com/travelset/model/repository/ItemRepository.java
@@ -1,10 +1,0 @@
-package com.travelset.model.repository;
-
-import com.travelset.model.entity.Item;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ItemRepository extends JpaRepository<Item, Integer> {
-
-}
