--- src/main/java/com/travelset/model/repository/PermissionRepository.java
@@ -1,10 +1,0 @@
-package com.travelset.model.repository;
-
-import com.travelset.model.entity.Permission;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PermissionRepository extends JpaRepository<Permission, Integer> {
-
-}
