--- src/main/java/com/travelset/model/repository/StatusRepository.java
@@ -1,10 +1,0 @@
-package com.travelset.model.repository;
-
-import com.travelset.model.entity.Status;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface StatusRepository extends JpaRepository<Status, Integer> {
-
-}
