--- src/main/java/com/travelset/model/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.travelset.model.repository;
-
-import com.travelset.model.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-}
