--- src/main/java/com/travelset/services/CategoryService.java
@@ -1,7 +1,0 @@
-package com.travelset.services;
-
-import com.travelset.model.entity.Category;
-
-public interface CategoryService {
-    Category getCategoryById(Integer id);
-}
