--- src/main/java/com/travelset/services/OrderlineService.java
@@ -1,5 +1,0 @@
-package com.travelset.services;
-
-public interface OrderlineService {
-    void createOrUpdateOrderline(int orderId, int itemId, int quantity);
-}
